feat: Add RS_X_InputDialogNoBlur addon plugin
This commit adds the RS_X_InputDialogNoBlur addon plugin, which allows the removal of the blur effect on the background of the input dialog. This plugin is targeted for RPG Maker MZ and is based on the RS_InputDialog plugin.

Author: biud436
URL: [https://github.com/biud436](https://github.com/biud436)

Change Log:
- 2024.10.24 (v1.0.0) - First Release.